#1918: ensure createDeltaPkg does not purge files that were just created in case of multiple source-target-maps
