+++ lib/Builder.js
@@ -293,6 +293,39 @@ saved
         }
         return responseObj;
     }
+    /**
+     * helper for buildDefinitionBulk, createDeltaPkg
+     *
+     * @param {string} listName market list name
+     * @returns {Promise.<void>} -
+     */
+    static async purgeDeployFolderList(listName) {
+        const properties = await config.getProperties();
+        for (const businessUnit in properties.marketList[listName]) {
+            if (businessUnit === 'description') {
+                // skip, it's just a metadata on this list and not a BU
+                continue;
+            }
+            await this.purgeDeployFolder(businessUnit);
+        }
+    }
+    /**
+     * helper for buildDefiniton, purgeDeployFolderList
+     *
+     * @param {string} businessUnit cred/bu combo
+     * @returns {Promise.<void>} -
+     */
+    static async purgeDeployFolder(businessUnit) {
+        const properties = await config.getProperties();
+        const deployDir = File.normalizePath([
+            properties.directories.deploy,
+            ...businessUnit.split('/'),
+        ]);
+        // Clear output folder structure for selected sub-type
+        // only run this if the standard deploy folder is a target of buildDefinition (which technically could be changed)
+        Util.logger.info(` - 🚮 purging folder ${deployDir}`);
+        await File.remove(deployDir);
+    }
 }
 
 export default Builder;
