+++ lib/index.js
@@ -1314,18 +1314,9 @@ class Mcdev {
             return;
         }
 
-        // standard deploy folder
-        const buObject = await Cli.getCredentialObject(properties, businessUnit);
-        const deployDir = File.normalizePath([
-            properties.directories.deploy,
-            buObject.credential,
-            buObject.businessUnit,
-        ]);
         if (Util.OPTIONS.purge) {
-            // Clear output folder structure for selected sub-type
-            // only run this if the standard deploy folder is a target of buildDefinition (which technically could be changed)
-            Util.logger.info(` - 🚮 purging folder ${deployDir}`);
-            await File.remove(deployDir);
+            const buObject = await Cli.getCredentialObject(properties, businessUnit);
+            Builder.purgeDeployFolder(buObject.credential + '/' + buObject.businessUnit);
         } else {
             Util.logger.info(` ☇ skipping purge of folder`);
         }
@@ -1374,20 +1365,7 @@ class Mcdev {
             return;
         }
         if (Util.OPTIONS.purge) {
-            for (const businessUnit in properties.marketList[listName]) {
-                if (businessUnit === 'description') {
-                    // skip, it's just a metadata on this list and not a BU
-                    continue;
-                }
-                const deployDir = File.normalizePath([
-                    properties.directories.deploy,
-                    ...businessUnit.split('/'),
-                ]);
-                // Clear output folder structure for selected sub-type
-                // only run this if the standard deploy folder is a target of buildDefinition (which technically could be changed)
-                Util.logger.info(` - 🚮 purging folder ${deployDir}`);
-                await File.remove(deployDir);
-            }
+            Builder.purgeDeployFolderList(listName);
         } else {
             Util.logger.info(` ☇ skipping purge of folder`);
         }
