+++ lib/util/devops.js
@@ -6,6 +6,7 @@ import Cli from './cli.js';
 import { simpleGit } from 'simple-git';
 const git = simpleGit();
 import mcdev from '../index.js';
+import Builder from '../Builder.js';
 import MetadataType from '../MetadataTypeInfo.js';
 import jsonToTable from 'json-to-table';
 import pLimit from 'p-limit';
@@ -448,6 +449,15 @@ const DevOps = {
                 Util.logger.error('Deployment Target: ' + ex.message);
             }
         }
+        if (Util.OPTIONS.purge === true && sourceMarketListArr.length > 1) {
+            // if --purge was defined and there is more than one source-target mapping, execute the purge up front to avoid deleting the package that was created by a prior mapping in this same run
+            for (const sourceMlName of sourceMarketListArr) {
+                /** @type {string} */
+                Builder.purgeDeployFolderList(sourceTargetMapping[sourceMlName]);
+            }
+            Util.OPTIONS.purge = false;
+        }
+
         // all good let's loop a second time for actual execution
         for (const sourceMlName of sourceMarketListArr) {
             /** @type {string} */
